create resource login

+++ resources/views/exports/index.blade.php
@@ -34,14 +34,17 @@
 						<td>{{ strtoupper($export->nama_pelabuhan) }}</td>
 						<td>{{ strtoupper($export->berat_bersih) }}</td>
 						<td>{{ strtoupper($export->nilai) }}</td>
-						<td class="col-md-1" align="right">
-							<a href="/exports/{{$export->id}}/edit" class="btn btn-xs btn-info">Edit</a> 
-						</td>
-						<td class="col-md-1 delete" align="left">
-							{!! Form::open(['method'=>'delete', 'route'=>['exports.destroy', $export->id]]) !!}
-							{!! Form::submit('Delete', ['class'=>'btn btn-xs btn-danger']) !!}
-							{!!Form::close()!!}
-						</td>
+
+            @if (!Auth::guest())
+							<td class="col-md-1" align="right">
+								<a href="/exports/{{$export->id}}/edit" class="btn btn-xs btn-info">Edit</a> 
+							</td>
+							<td class="col-md-1 delete" align="left">
+								{!! Form::open(['method'=>'delete', 'route'=>['exports.destroy', $export->id]]) !!}
+								{!! Form::submit('Delete', ['class'=>'btn btn-xs btn-danger']) !!}
+								{!!Form::close()!!}
+							</td>
+						@endif
 					</tr>
 				<?php $i += 1 ?>
 				@endforeach
